goal db call refactoring + sort goals properly
moved widget to correct screen
created function for db call used twice
sorted goals properly on dashboard
